Removed logic from EvoAgent benchmark class
