+++ evo_researcher/main.py
@@ -1,8 +1,9 @@
 import click
 import time
 from dotenv import load_dotenv
-from evo_researcher.benchmark.agents import EvoAgent
+from evo_researcher.benchmark.agents import _make_prediction
 from langchain_community.callbacks import get_openai_callback
+from evo_researcher.functions.research import research as evo_research
 
 load_dotenv()
 
@@ -30,11 +31,10 @@ def research(
     prompt: str,
     file: str | None = None
 ) -> None:
-    agent = EvoAgent(model="gpt-4-1106-preview")
     start = time.time()
     
     with get_openai_callback() as cb:
-      report = agent.research(goal=prompt, use_summaries=False)
+      report = evo_research(goal=prompt, use_summaries=False, model="gpt-4-1106-preview")
     
     end = time.time()
     
@@ -52,16 +52,15 @@ def research(
 @click.argument('prompt')
 @click.option('--path', '-p', default=None)
 def predict(prompt: str, path: str | None = None) -> None:
-    agent = EvoAgent(model="gpt-4-1106-preview")
     start = time.time()
 
     with get_openai_callback() as cb:
         if path:
             report = read_text_file(path)
         else:
-            report = agent.research(goal=prompt, model="gpt-4-1106-preview", use_summaries=False)
+            report = evo_research(goal=prompt, model="gpt-4-1106-preview", use_summaries=False)
         
-        prediction = agent.predict_from_research(market_question=prompt, research_report=report)
+        prediction = _make_prediction(market_question=prompt, additional_information=report, engine="gpt-4-1106-preview", temperature=0.0)
 
     end = time.time()
